+++ Runtime/BaseModule/Constructables/IConstructable.cs
@@ -3,7 +3,7 @@ namespace FlowIoC.BaseModule.Constructables
     public interface IConstructable
     {
         bool IsPostConstructed { get; set; }
-        bool IsDeConstructed { get; set; }
+        bool IsDeconstructed { get; set; }
         void PostConstruct();
         void Deconstruct() { }
     }