+++ Runtime/ScreenModule/Model/Runtime/ScreenRuntimeModel.cs
@@ -21,7 +21,7 @@ namespace FlowIoC.ScreenModule.Model.Runtime
         private Transform _poolParent;
 
         public bool IsPostConstructed { get; set; }
-        public bool IsDeConstructed { get; set; }
+        public bool IsDeconstructed { get; set; }
 
         public void PostConstruct()
         {